+++ PerfectPronunciation/Model/Leaderboard.swift
@@ -11,6 +11,6 @@ struct Leaderboard: Identifiable {
     var id : String
     var userName : String
     var country : String
-    var weeklyChallengeComplete : Float
+    var weeklyChallengeComplete : Double
     
 }
